chore(webpack): consolidate most config into common file
Most of the Webpack configuration is the same between Dev and Prod.
With configuration spread across three files, it was difficult to reason about
which plugins were being used where, and in what ways.

This change keeps the existing file structure and build scripts in place,
but reduces the scope of webpack.dev and webpack.prod to only set the 
environment variables, and load the complete config from webpack.common.

All plugins and configuration exist in webpack.common, with some env-specific
conditions to modify the final configuration object.