updated metadata schema specification with changes
The following changes aim to clarify the requirements for compliance with
version 1.0.1 of the code.gov metadata schema:

Schema version: The first official iteration of the the code.gov metadata
schema will be version 1.0.1. Adding a version number to the schema makes
it easier to track and manage changes between versions of the metadata
schema.

‘Organization’ field: The ‘organization’ field, which holds the name of the
sub-agency responsible for a particular repo or project, will now (1) be an
optional field and (2) move into the ‘projects’ object. This enables
agencies to identify the appropriate organization within that agency that
owns the repository, if such an organization exists.

‘Projects’ field: The code.gov policy guide previously referred to a
‘project’ field in the metadata schema, while the sample code.json referred
to the plural from ‘projects’. The policy guide will be updated to refer to
the plural form ‘projects’.

‘govwideReuseproject’ field: The code.gov policy guide previously referred
to ‘governmentWideReuseProject’ field in the metadata schema, while the
sample code.json referred to ‘govwideReuseproject’. The sample code.json
will be updated to refer to ‘‘governmentWideReuseProject’ in order to
conform with the code.gov policy guide.

‘tags’ field: The code.gov policy guide and the sample json refer to a
‘tags’ field, defined as an array of strings. However, previous iterations
of the metadata schema and the sample code referred to a ‘projectTags’
field, which was structured as an array of ‘tag’ objects. To clarify, for
version 1.0.1, we intend to define the 'tags’ field as an array of strings
for simplicity and consistency with the ‘languages’ field.

‘exemption_text’ field: This field allows agencies to provide a brief
narrative explanation for the exception requested in the 'exemption' field.